projects/coach-amechi/e_commerce/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-resized>=1.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faker>=38.0.0",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
